out of statistical equilibrium

+++ simplecosmomodels.h
@@ -17,10 +17,11 @@ void derivs2(double t, double*y, double*dydt);
 void desitter();
 void deLCDM();
 void einsteindesitter();
-double reciprocal_coef(double Tkin, double Eul, int gl, int gu, double coef_ul);
+double reciprocal_coef(double Tkin, double Eul, double gl, double gu, double coef_ul);
 double expansion(double t, double a);
 double BoltzmanH2ROP(double T);
 int expansion_calc(double **t, double **a, double **Trad,double **Tb);
+
